Update README.md
C-Language-beginning-to-advance


C-Language-beginning-to-advance helps plan, test, and automate your migration to C-Language-beginning-to-advance from the following platforms:
•	C language
How to request support
If you need assistance, you can file a support ticket here.
Getting started
C-Language-beginning-to-advance is distributed as a kh7609528 and this extension to the official GitHub CLI to interact with the kh7609528.
Prerequisites
The following requirements must be met to be able to use the C-Language-beginning-to-advance:
•	The official GitHub CLI must be installed.
•	You must have credentials to authenticate with the GitHub Container Registry.
Installation
Next, C-Language-beginning-to-advance can be installed via this command:
$ gh extension install kh7609528/C-Language-beginning-to-advance
Configuration
New versions of the C-Language-beginning-to-advance Importer are released on a regular basis. To ensure you're up to date, run the following command:
$ C-Language-beginning-to-advance update
You can find detailed information about using environment variables in the platform-specific documentation..
How to offer feedback or make a feature request
If you would like to offer feedback or make a feature request, please create a new discussion Here.